+++ src/components/_share/JinniProvider/index.ts
@@ -1,3 +1,4 @@
 export { default as JinniProvider } from './JinniProvider';
-export type { DesignSystemType } from './JinniProvider';
+export type { DesignSystemType } from './JinniProvider.types';
 export { DEFAULT_DESIGN_SYSTEM } from './JinniProvider.constants';
+export { createDesignSystem } from './createDesignSystem';
